+++ jpos/src/main/java/org/jpos/iso/IVA_ALPHANUMNOBLANK.java
@@ -80,7 +80,7 @@ public class IVA_ALPHANUMNOBLANK extends IVA_ALPHANUM {
                 else
                     c = new ISOVField( c, e );
                 if ( breakOnError )
-                    throw new ISOVException ( "Error on field " + (Integer) c.getKey(), c );
+                    throw new ISOVException ( "Error on field " + c.getKey(), c );
             }
             return c;
         }
