+++ jpos/src/main/java/org/jpos/iso/IVA_ALPHANUMNOZERO.java
@@ -79,7 +79,7 @@ public class IVA_ALPHANUMNOZERO extends IVA_ALPHANUM {
                 else
                     c = new ISOVField( c, e );
                 if ( breakOnError )
-                    throw new ISOVException ( "Error on field " + (Integer) c.getKey(), c );
+                    throw new ISOVException ( "Error on field " + c.getKey(), c );
             }
             return c;
         }
